// src/app/admin_adminpersonal/interfaces/personal.interface.ts
export interface Personal{
  usuario_id: number;
  usuario_carnet_identidad: string;
  usuario_tipo_carnet: string;
  apellidos_nombres: string;
  usuario_direccion: string;
  usuario_fecha_nacimiento: string;
  usuario_telefono: string;
  usuario_correo: string;
  usuario_genero: string;
  usuario_nombre_usuario: string;
  usuario_contrasena: string;
}